// src/lib/constants.d.ts
export declare const DESTINATIONS: {
    name: string;
    slug: string;
    location: string;
    price: string;
    rating: number;
    image: string;
    duration: string;
    type: string;
}[];
export declare const DOMESTIC_DESTINATIONS: {
    name: string;
    slug: string;
    location: string;
    price: string;
    duration: string;
    rating: number;
    image: string;
    type: string;
}[];
export declare const TRAVEL_STORIES: {
    image: string;
    likes: number;
    comments: number;
}[];
export declare const HERO_IMAGES: {
    main: string;
    service1: string;
    service2: string;
    service3: string;
    whyChooseUs1: string;
    whyChooseUs2: string;
};